Consolidate the three implementations of Decode5A3 found in Core into one. Fix some warnings.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2364 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/StringUtil.cpp
@@ -76,7 +76,7 @@ u32 Ascii2Hex(std::string _Text)
 	u32 Result = 0;
 
 	// Max 32-bit values are supported
-	int Length = _Text.length();
+	size_t Length = _Text.length();
 	if (Length > 4)
 		Length = 4;
